requirements.txt
asgiref==3.2.10
dj-database-url==0.5.0
Django==3.0.7
django-cors-headers==3.4.0
django-db-constraints==0.3.0
django-extra-fields==2.0.5
djangorestframework==3.11.0
gunicorn==20.0.4
postgis==1.0.4
psycopg2==2.8.5
pytz==2020.1
sqlparse==0.3.1
whitenoise==5.0.1
